requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
gunicorn
